+++ lib/handler/raw_test.go
@@ -398,7 +398,7 @@ func TestParseSendPort_Raw(t *testing.T) {
 	tests := []struct {
 		input   string
 		name    string
-		want    int
+		want    uint16
 		wantErr bool
 	}{
 		{"0", "PORT", 0, false},
